Job details page

+++ db/migrate/20191217222108_create_jobs.rb
@@ -4,7 +4,7 @@ class CreateJobs < ActiveRecord::Migration[5.2]
       t.string :title
       t.text :description
       t.text :skills
-      t.integer :salary
+      t.float :salary
       t.integer :job_level
       t.date :start_date
       t.date :end_date
